refactor: move common util functions
